pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "the-ciso-council"
version = "0.1.0"
description = "A virtual council of AI models debating real GRC and cybersecurity dilemmas"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Kunal RK", email = "kunalrk@example.com"},
]
keywords = ["grc", "cybersecurity", "llm", "ai", "governance", "risk", "compliance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

dependencies = [
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
ciso-council = "council.cli:main"

[project.urls]
Homepage = "https://github.com/kunalrk/the-ciso-council"
Repository = "https://github.com/kunalrk/the-ciso-council"
Issues = "https://github.com/kunalrk/the-ciso-council/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
